requirements.txt
requests
json
re
bidi.algorithm 
arabic_reshaper
selenium
bs4